test(arm64): widen bulk_memory tests; dedup passiveDataModule
Move passiveDataModule (previously duplicated between bulk_memory_test.go and the
darwin&&arm64 helper) to a broadly-tagged dataseg_shared_test.go, then widen
bulk_memory_test.go to arm64. Passes on darwin/arm64 and linux/amd64.
